Unify plugin app registration collection
