+++ packages/plugin-sdk/src/testing/app.tsx
@@ -56,18 +56,8 @@ import {
   type JsonValue,
 } from "@get-bb/plugin-sdk";
 import { isComposerDraftEmpty } from "../internal/composer-view.js";
-import {
-  collectComposerCustomization,
-  normalizePluginThreadRowStatus,
-  PLUGIN_SLOT_ID_PATTERN,
-  requireComponent,
-  requireMessageDirectiveId,
-  requireNonEmptyString,
-  requireOptionalString,
-  requireProviderId,
-  requireSlotId,
-  requireUniqueId,
-} from "../internal/composer-customization-validation.js";
+import { normalizePluginThreadRowStatus } from "../internal/composer-customization-validation.js";
+import { collectPluginAppRegistrations } from "../internal/plugin-app-collector.js";
 
 /**
  * `@get-bb/plugin-sdk/testing/app` — the frontend plugin test harness. Tests a
@@ -102,7 +92,6 @@ export interface RpcCall {
   method: string;
   input: unknown;
 }
-
 export type NavigateCall =
   | { method: "toThread"; threadId: string }
   | { method: "toProject"; projectId: string }
@@ -540,317 +529,6 @@ export type PluginAppSource =
   | (() => Promise<PluginAppDefinition | PluginAppModule>);
 
 /**
- * Uses the same registration validation as the BB app collector so a
- * registration the host would reject fails here with the same message.
- */
-function collectRegistrations(
-  definition: PluginAppDefinition,
-): CapturedPluginApp {
-  const captured: CapturedPluginApp = {
-    homepageSections: [],
-    settingsSections: [],
-    navPanels: [],
-    threadPanelActions: [],
-    newThreadPanelActions: [],
-    composerCustomizations: [],
-    pendingInteractions: [],
-    sidebarFooterActions: [],
-    threadLists: [],
-    threadHeaderActions: [],
-    fileOpeners: [],
-    messageDirectives: [],
-    messageActions: [],
-    providerIcons: [],
-    contentScripts: [],
-  };
-  const seenIds = {
-    homepageSection: new Set<string>(),
-    settingsSection: new Set<string>(),
-    navPanel: new Set<string>(),
-    threadPanelAction: new Set<string>(),
-    newThreadPanelAction: new Set<string>(),
-    composerCustomization: new Set<string>(),
-    pendingInteraction: new Set<string>(),
-    sidebarFooterAction: new Set<string>(),
-    threadList: new Set<string>(),
-    threadHeaderAction: new Set<string>(),
-    fileOpener: new Set<string>(),
-    messageDirective: new Set<string>(),
-    messageAction: new Set<string>(),
-    providerIcon: new Set<string>(),
-    contentScript: new Set<string>(),
-  };
-
-  definition.setup({
-    slots: {
-      homepageSection(registration) {
-        const kind = "slots.homepageSection";
-        const id = requireSlotId(kind, registration?.id);
-        requireUniqueId(kind, seenIds.homepageSection, id);
-        captured.homepageSections.push({
-          id,
-          title: requireNonEmptyString(kind, "title", registration.title),
-          component: requireComponent(kind, registration.component),
-        });
-      },
-      settingsSection(registration) {
-        const kind = "slots.settingsSection";
-        const id = requireSlotId(kind, registration?.id);
-        requireUniqueId(kind, seenIds.settingsSection, id);
-        const title = requireOptionalString(kind, "title", registration.title);
-        const description = requireOptionalString(
-          kind,
-          "description",
-          registration.description,
-        );
-        captured.settingsSections.push({
-          id,
-          ...(title !== undefined ? { title } : {}),
-          ...(description !== undefined ? { description } : {}),
-          component: requireComponent(kind, registration.component),
-        });
-      },
-      navPanel(registration) {
-        const kind = "slots.navPanel";
-        const id = requireSlotId(kind, registration?.id);
-        requireUniqueId(kind, seenIds.navPanel, id);
-        const path = requireNonEmptyString(kind, "path", registration.path);
-        if (!PLUGIN_SLOT_ID_PATTERN.test(path)) {
-          throw new Error(
-            `${kind}: "path" must match ${String(PLUGIN_SLOT_ID_PATTERN)} (it becomes a URL segment), got ${JSON.stringify(path)}`,
-          );
-        }
-        if (
-          registration.headerContent !== undefined &&
-          typeof registration.headerContent !== "function"
-        ) {
-          throw new Error(
-            `${kind}: "headerContent" must be a React component function when set`,
-          );
-        }
-        if (
-          registration.experimental_sidebarAccessory !== undefined &&
-          typeof registration.experimental_sidebarAccessory !== "function"
-        ) {
-          throw new Error(
-            `${kind}: "experimental_sidebarAccessory" must be a React component function when set`,
-          );
-        }
-        captured.navPanels.push({
-          id,
-          title: requireNonEmptyString(kind, "title", registration.title),
-          icon: requireNonEmptyString(kind, "icon", registration.icon),
-          path,
-          component: requireComponent(kind, registration.component),
-          ...(registration.experimental_sidebarAccessory !== undefined
-            ? {
-                experimental_sidebarAccessory:
-                  registration.experimental_sidebarAccessory,
-              }
-            : {}),
-          ...(registration.headerContent !== undefined
-            ? { headerContent: registration.headerContent }
-            : {}),
-        });
-      },
-      threadPanelAction(registration) {
-        const kind = "slots.threadPanelAction";
-        const id = requireSlotId(kind, registration?.id);
-        requireUniqueId(kind, seenIds.threadPanelAction, id);
-        if (
-          registration.run !== undefined &&
-          typeof registration.run !== "function"
-        ) {
-          throw new Error(`${kind}: "run" must be a function when set`);
-        }
-        if (
-          registration.layout !== undefined &&
-          registration.layout !== "padded" &&
-          registration.layout !== "flush"
-        ) {
-          throw new Error(`${kind}: "layout" must be "padded" or "flush"`);
-        }
-        captured.threadPanelActions.push({
-          id,
-          title: requireNonEmptyString(kind, "title", registration.title),
-          ...(registration.icon !== undefined
-            ? { icon: requireNonEmptyString(kind, "icon", registration.icon) }
-            : {}),
-          component: requireComponent(kind, registration.component),
-          ...(registration.layout !== undefined
-            ? { layout: registration.layout }
-            : {}),
-          ...(registration.run !== undefined ? { run: registration.run } : {}),
-        });
-      },
-      experimental_newThreadPanelAction(registration) {
-        const kind = "slots.experimental_newThreadPanelAction";
-        const id = requireSlotId(kind, registration?.id);
-        requireUniqueId(kind, seenIds.newThreadPanelAction, id);
-        if (
-          registration.run !== undefined &&
-          typeof registration.run !== "function"
-        ) {
-          throw new Error(`${kind}: "run" must be a function when set`);
-        }
-        if (
-          registration.layout !== undefined &&
-          registration.layout !== "padded" &&
-          registration.layout !== "flush"
-        ) {
-          throw new Error(`${kind}: "layout" must be "padded" or "flush"`);
-        }
-        captured.newThreadPanelActions.push({
-          id,
-          title: requireNonEmptyString(kind, "title", registration.title),
-          ...(registration.icon !== undefined
-            ? { icon: requireNonEmptyString(kind, "icon", registration.icon) }
-            : {}),
-          component: requireComponent(kind, registration.component),
-          ...(registration.layout !== undefined
-            ? { layout: registration.layout }
-            : {}),
-          ...(registration.run !== undefined ? { run: registration.run } : {}),
-        });
-      },
-      pendingInteraction(registration) {
-        const kind = "slots.pendingInteraction";
-        const id = requireSlotId(kind, registration?.id);
-        requireUniqueId(kind, seenIds.pendingInteraction, id);
-        captured.pendingInteractions.push({
-          id,
-          component: requireComponent(kind, registration.component),
-        });
-      },
-      sidebarFooterAction(registration) {
-        const kind = "slots.sidebarFooterAction";
-        const id = requireSlotId(kind, registration?.id);
-        requireUniqueId(kind, seenIds.sidebarFooterAction, id);
-        if (typeof registration.run !== "function") {
-          throw new Error(`${kind}: "run" must be a function`);
-        }
-        captured.sidebarFooterActions.push({
-          id,
-          title: requireNonEmptyString(kind, "title", registration.title),
-          icon: requireNonEmptyString(kind, "icon", registration.icon),
-          run: registration.run,
-        });
-      },
-      experimental_threadList(registration) {
-        const kind = "slots.experimental_threadList";
-        const id = requireSlotId(kind, registration?.id);
-        requireUniqueId(kind, seenIds.threadList, id);
-        const description = requireOptionalString(
-          kind,
-          "description",
-          registration.description,
-        );
-        captured.threadLists.push({
-          id,
-          title: requireNonEmptyString(kind, "title", registration.title),
-          ...(description !== undefined ? { description } : {}),
-          component: requireComponent(kind, registration.component),
-        });
-      },
-      experimental_threadHeaderAction(registration) {
-        const kind = "slots.experimental_threadHeaderAction";
-        const id = requireSlotId(kind, registration?.id);
-        requireUniqueId(kind, seenIds.threadHeaderAction, id);
-        captured.threadHeaderActions.push({
-          id,
-          title: requireNonEmptyString(kind, "title", registration.title),
-          component: requireComponent(kind, registration.component),
-        });
-      },
-      fileOpener(registration) {
-        const kind = "slots.fileOpener";
-        const id = requireSlotId(kind, registration?.id);
-        requireUniqueId(kind, seenIds.fileOpener, id);
-        const rawExtensions = registration?.extensions;
-        if (!Array.isArray(rawExtensions) || rawExtensions.length === 0) {
-          throw new Error(
-            `${kind}: "extensions" must be a non-empty array of lowercase extensions without the dot`,
-          );
-        }
-        const extensions = rawExtensions.map((extension) => {
-          if (typeof extension !== "string" || !/^[a-z0-9]+$/.test(extension)) {
-            throw new Error(
-              `${kind}: extensions must be lowercase alphanumerics without the dot, got ${JSON.stringify(extension)}`,
-            );
-          }
-          return extension;
-        });
-        captured.fileOpeners.push({
-          id,
-          title: requireNonEmptyString(kind, "title", registration.title),
-          extensions,
-          component: requireComponent(kind, registration.component),
-        });
-      },
-      messageDirective(registration) {
-        const kind = "slots.messageDirective";
-        const id = requireMessageDirectiveId(kind, registration?.id);
-        requireUniqueId(kind, seenIds.messageDirective, id);
-        captured.messageDirectives.push({
-          id,
-          component: requireComponent(kind, registration.component),
-        });
-      },
-      messageAction(registration) {
-        const kind = "slots.messageAction";
-        const id = requireSlotId(kind, registration?.id);
-        requireUniqueId(kind, seenIds.messageAction, id);
-        if (typeof registration.run !== "function") {
-          throw new Error(`${kind}: "run" must be a function`);
-        }
-        captured.messageActions.push({
-          id,
-          title: requireNonEmptyString(kind, "title", registration.title),
-          ...(registration.icon !== undefined
-            ? { icon: requireNonEmptyString(kind, "icon", registration.icon) }
-            : {}),
-          run: registration.run,
-        });
-      },
-      experimental_providerIcon(registration) {
-        const kind = "slots.experimental_providerIcon";
-        const providerId = requireProviderId(kind, registration?.providerId);
-        requireUniqueId(kind, seenIds.providerIcon, providerId);
-        captured.providerIcons.push({
-          providerId,
-          icon: requireComponent(kind, registration.icon),
-        });
-      },
-    },
-    composer: {
-      customize(registration) {
-        const customization = collectComposerCustomization(
-          registration,
-          seenIds.composerCustomization,
-          (reason) => console.warn(reason),
-        );
-        if (customization !== null) {
-          captured.composerCustomizations.push(customization);
-        }
-      },
-    },
-    contentScripts: {
-      register(registration) {
-        const kind = "contentScripts.register";
-        const id = requireSlotId(kind, registration?.id);
-        requireUniqueId(kind, seenIds.contentScript, id);
-        if (typeof registration.mount !== "function") {
-          throw new Error(`${kind}: "mount" must be a function`);
-        }
-        captured.contentScripts.push({ id, mount: registration.mount });
-      },
-    },
-  });
-
-  return captured;
-}
-
-/**
  * Install the test runtime, resolve the plugin app definition, and capture
  * its slot registrations. Pass a thunk (`() => import("../app.tsx")`) so the
  * plugin module evaluates after the runtime is installed — a static import
@@ -869,7 +547,7 @@ export async function loadPluginApp(
       "the bundle's default export is not definePluginApp(...) from @get-bb/plugin-sdk/app",
     );
   }
-  return collectRegistrations(definition);
+  return collectPluginAppRegistrations(definition);
 }
 
 export interface ContentScriptTestMountOptions {
